% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangertoolkit-package.R
\docType{package}
\name{rangertoolkit-package}
\alias{rangertoolkit}
\alias{rangertoolkit-package}
\title{rangertoolkit: Utilities for Ranger Random Forest Models}
\description{
Convenience functions to make working with Ranger random forest models easier.
}
\author{
\strong{Maintainer}: Roberts Nelson \email{rnelsonxy@gmail.com} (\href{https://orcid.org/0000-0003-4644-8380}{ORCID})

}
\keyword{internal}
