% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_oob.R
\name{score_oob}
\alias{score_oob}
\title{Score out-of-bag samples}
\usage{
score_oob(model, X, y)
}
\arguments{
\item{model}{A ranger random forest model.}

\item{X}{Training data feature matrix.}

\item{y}{Training data response vector.}
}
\value{
A tibble summarizing OOB performance.
}
\description{
Score out-of-bag samples
}
\examples{

# regression

regression_model <- ranger::ranger(
  mpg ~ .,
  data = mtcars,
  num.trees = 100,
  keep.inbag = TRUE
)
score_oob(
  regression_model,
  mtcars[, -1],
  mtcars[, 1]
)

# survival

lung_clean <- na.omit(survival::lung)
rsf <- ranger::ranger(
  survival::Surv(time, status) ~ .,
  data = lung_clean,
  num.trees = 100,
  keep.inbag=TRUE
)
score_oob(
  rsf,
  lung_clean[, -c(2, 3)],
  survival::Surv(lung_clean$time, lung_clean$status)
)
}
